refactor: move code to script file
